Added Required Repositories

+++ src/main/java/com/WebApplication/Model/OrderItems.java
@@ -2,12 +2,13 @@ package com.WebApplication.Model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public class OrderItems {
     @Id
-    private int orderId;
+    private Long orderId;
     private int quantity;
-    private float price;
+    private BigDecimal price;
 
 }
