+++ src/main/java/com/WebApplication/Model/Products.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,10 +14,10 @@ import lombok.NoArgsConstructor;
 public class Products {
 
     @Id
-    private int productId;
+    private Long productId;
     private String productName;
     private  String description;
-    private int price;
+    private BigDecimal price;
     private int stackQuantity;
 
 }
